requirements.txt

langchain_core
python-dotenv
streamlit
langchain_community
langserve
faiss-cpu
groq
langchain-groq
langchainhub
sentence_transformers
langchain-objectbox
langchain
torch
torchvision
torchaudio
tensorflow 
tensorflow-hub